+++ libs/interpreter-lib/src/index.ts
@@ -3,6 +3,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
 export * from './interpreter';
-export * from './cli-util';
+export * from './parsing-util';
 export * from './logging';
 export * from './validation-checks';
